Fix : 키값 수정

+++ back/src/main/java/com/back/domain/mentoring/session/repository/MentoringSessionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MentoringSessionRepository extends JpaRepository<MentoringSession, Integer> {
+public interface MentoringSessionRepository extends JpaRepository<MentoringSession, Long> {
 }
